+++ History.md
@@ -1,3 +1,13 @@
+# 1.5.1-SNAPSHOT / ????-??-??
+
+## New and Noteworthy
+
+## Enhancements
+
+## Fixed Issues
+
+## External Contributions
+
 # 1.5.0 / 2022-05-06
 
 ## Enhancements